feat(S2): 게임 플레이 버튼 분리
