+++ apps/game-builder/src/components/game/intro/GameIntro.tsx
@@ -1,14 +1,14 @@
 "use client";
-import Link from "next/link";
-import { useRouter } from "next/navigation";
 import { ClockIcon, InfoCircledIcon, Pencil1Icon } from "@radix-ui/react-icons";
 import { AspectRatio } from "@repo/ui/components/ui/AspectRatio";
 import type { GameIntro as GameIntroType } from "@/interface/customType";
 import DateDisplay from "@/components/common/text/DateDisplay";
 import { DynamicViewer } from "@/components/common/viewer/DynamicViewer";
-import { Button } from "@/packages/ui/components/ui/Button";
 import TextWithCounts from "@/components/common/text/TextWithCounts";
 import TextWithNumberRange from "@/components/common/text/TextWithCountsRange";
+import GameRestartButton from "@/components/button/GameRestartButton";
+import GameContinueButton from "@/components/button/GameContinueButton";
+import GameStartButton from "@/components/button/GameStartButton";
 
 export default function GameIntro({
   gameIntroData,
@@ -17,17 +17,10 @@ export default function GameIntro({
   gameIntroData: GameIntroType;
   gameId: number;
 }) {
-  const router = useRouter();
   const gameData = gameIntroData.game;
   const subData = gameIntroData.enrichData;
   // FIXME: api에서 받은 값으로 변경할 것
-  const isFirstStart = gameIntroData.enrichData;
-
-  const handleRestartClick = () => {
-    if (confirm("게임을 처음부터 시작하시겠습니까?")) {
-      router.push(`/game/${gameId}/play/?play=first`);
-    }
-  };
+  const playId = 1;
 
   return (
     <section className="pt-10 text-center">
@@ -80,31 +73,13 @@ export default function GameIntro({
         </div>
         <div className="xs:mx-6 sm:mx-12 my-6 flex flex-col gap-4">
           <div className="text-xs">마지막 플레이한 페이지 요약</div>
-          {!isFirstStart ? (
+          {playId === undefined ? (
             <>
-              <Link
-                href={`gameId/${gameId}/play/${gameId}?play=continue`}
-                className="w-full"
-              >
-                <Button className="w-full">이어하기</Button>
-              </Link>
-              <Button
-                className="w-full"
-                variant="outline"
-                onClick={handleRestartClick}
-              >
-                새로 하기
-              </Button>
+              <GameContinueButton gameId={gameId} playId={playId} />
+              <GameRestartButton gameId={gameId} />
             </>
           ) : (
-            <Link
-              href={`gameId/${gameId}/play/${gameId}?play=first-start`}
-              className="w-full"
-            >
-              <Button className="w-full" variant="outline">
-                게임 시작
-              </Button>
-            </Link>
+            <GameStartButton gameId={gameId} />
           )}
         </div>
       </div>
